save picture bug fix.

+++ grab.py
@@ -4,6 +4,9 @@ import os
 import re
 
 
+url_home = 'http://2479.com/'
+
+
 def request(url, name):
     if os.path.exists(name):
         print 'ignore %s' % url
@@ -12,6 +15,19 @@ def request(url, name):
     print 'downloading %s' % url
     r = requests.get(url)
     content = r.text.encode(r.encoding or 'utf8')
+    with open(name, 'wb') as f:
+        f.write(content)
+    return content
+
+
+def request_img(url, name):
+    if os.path.exists(name):
+        print 'ignore %s' % url
+        return
+
+    print 'downloading %s' % url
+    r = requests.get(url)
+    content = r.content
     with open(name, 'wb') as f:
         f.write(content)
     return content
@@ -26,17 +42,47 @@ def download_page(url, name):
     links = find_css(c) + find_js(c)
 
     for link in links:
-        link_url = os.path.join(url, link)
+        try:
+            if link.startswith('/'):
+                link = link[1:]
+                link_url = os.path.join(url_home, link)
+            else:
+                link_url = os.path.join(url, link)
 
-        name = os.path.join('public', link)
-        if name.find('?') != -1:
-            name = name[:name.find('?')]
-        dirname = os.path.dirname(name)
-        if not os.path.exists(dirname):
-            print 'mkdir -p %s' % dirname
-            os.makedirs(dirname)
+            name = os.path.join('public', link)
+            if name.find('?') != -1:
+                name = name[:name.find('?')]
+            dirname = os.path.dirname(name)
+            if not os.path.exists(dirname):
+                print 'mkdir -p %s' % dirname
+                os.makedirs(dirname)
+            request(link_url, name)
+        except Exception as e:
+            print e
+            print link
 
-        request(link_url, name)
+    links = find_img(c)
+    for link in links:
+        try:
+            if link.startswith('/'):
+                link = link[1:]
+                link_url = os.path.join(url_home, link)
+            else:
+                link_url = os.path.join(url, link)
+
+            name = os.path.join('public', link)
+            if name.find('?') != -1:
+                name = name[:name.find('?')]
+            dirname = os.path.dirname(name)
+            if not os.path.exists(dirname):
+                print 'mkdir -p %s' % dirname
+                os.makedirs(dirname)
+            request_img(link_url, name)
+        except Exception as e:
+            print e
+            print link
+
+
 
 
 def find_css(content):
@@ -49,8 +95,12 @@ def find_js(content):
     return ptn.findall(content)
 
 
+def find_img(content):
+    ptn = re.compile(r'<img [^>]*src="([^"]+)"[^>]*>')
+    return ptn.findall(content)
+
+
 if __name__ == '__main__':
-    url_home = 'http://2479.com/'
     path = 'public'
     if not os.path.exists(path):
         os.makedirs(path)
